pyproject.toml
[project]
name = "curryparty"
version = "0.2.2"
description = "Python playground to learn lambda calculus"
readme = "README.md"
authors = [
    { name = "Antonin P", email = "antonin.peronnet@telecom-paris.fr" }
]
requires-python = ">=3.13"
dependencies = [
    "svg-py>=1.9.2",
]

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "marimo>=0.18.4",
    "polars>=1.36.1",
    "ruff>=0.14.8",
]
